requirements.txt
# 1. install torch .. 2. install requirements.txt 3. install pytorch3d flash attention 4. python build.py install in nndistance
# pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu117
# pip install git+https://github.com/Dao-AILab/flash-attention.git
scikit-image
albumentations
einops==0.3.0
# flash_attn==2.2.3.post2
h5py==3.1.0
hydra-core==1.3.2
kornia==0.6.11
loguru==0.5.3
matplotlib==3.7.3
mesh2sdf==1.1.0
natsort==8.4.0
numpy==1.25.2
omegaconf==2.3.0
open3d==0.17.0
opencv_python_headless==3.4.18.65
Pillow==10.0.0
pycolmap==0.3.0
pymeshlab==2021.7
pyrender==0.1.45
# pytorch3d==0.7.4
ray==1.13.0
rich==13.6.0
scikit_learn==1.3.0
scipy==1.11.3
# setuptools==68.2.1
timm==0.9.6
tqdm==4.66.1
trimesh==3.23.5
ujson==5.8.0
wandb==0.15.9
yacs==0.1.8
git+https://github.com/Dao-AILab/flash-attention.git
pytorch_lightning==1.5.0
